Silence some warnings, fix volume increments.

+++ src/Systems/BYTEPUSHER/Cores/BYTEPUSHER_STANDARD.hpp
@@ -13,13 +13,14 @@
 /*==================================================================*/
 
 class BYTEPUSHER_STANDARD final : public BytePusher_CoreInterface {
-	static constexpr u32 cTotalMemory{ MiB(16) };
-	static constexpr u32 cSafezoneOOB{      8 };
-	static constexpr f32 cRefreshRate{ 60.00f };
-	static constexpr s32 cAudioLength{    256 };
-	static constexpr s32 cResSizeMult{      2 };
-	static constexpr s32 cScreenSizeX{    256 };
-	static constexpr s32 cScreenSizeY{    256 };
+	static constexpr u64 cTotalMemory{ MiB(16) };
+	static constexpr u32 cSafezoneOOB{     8 };
+	static constexpr f32 cRefreshRate{ 60.0f };
+
+	static constexpr s32 cAudioLength{ 256 };
+	static constexpr s32 cResSizeMult{   2 };
+	static constexpr s32 cScreenSizeX{ 256 };
+	static constexpr s32 cScreenSizeY{ 256 };
 
 	static constexpr u32 cMaxDisplayW{ 256 };
 	static constexpr u32 cMaxDisplayH{ 256 };
