+++ src/Systems/CHIP8/Cores/MEGACHIP.hpp
@@ -15,22 +15,22 @@
 /*==================================================================*/
 
 class MEGACHIP final : public Chip8_CoreInterface {
-	static constexpr u32 cTotalMemory{ MiB(16) };
-	static constexpr u32 cSafezoneOOB{       32 };
-	static constexpr u32 cGameLoadPos{      512 };
-	static constexpr u32 cStartOffset{      512 };
-	static constexpr f32 cRefreshRate{    60.0f };
+	static constexpr u64 cTotalMemory{ MiB(16) };
+	static constexpr u32 cSafezoneOOB{    32 };
+	static constexpr u32 cGameLoadPos{   512 };
+	static constexpr u32 cStartOffset{   512 };
+	static constexpr f32 cRefreshRate{ 60.0f };
 
-	static constexpr s32 cResSizeMult{        4 };
-	static constexpr s32 cScreenSizeX{      128 };
-	static constexpr s32 cScreenSizeY{       64 };
-	static constexpr s32 cInstSpeedHi{       45 };
-	static constexpr s32 cInstSpeedLo{       30 };
+	static constexpr s32 cResSizeMult{   4 };
+	static constexpr s32 cScreenSizeX{ 128 };
+	static constexpr s32 cScreenSizeY{  64 };
+	static constexpr s32 cInstSpeedHi{  45 };
+	static constexpr s32 cInstSpeedLo{  30 };
 
-	static constexpr s32 cResMegaMult{        2 };
-	static constexpr s32 cScreenMegaX{      256 };
-	static constexpr s32 cScreenMegaY{      192 };
-	static constexpr s32 cInstSpeedMC{     3000 };
+	static constexpr s32 cResMegaMult{    2 };
+	static constexpr s32 cScreenMegaX{  256 };
+	static constexpr s32 cScreenMegaY{  192 };
+	static constexpr s32 cInstSpeedMC{ 3000 };
 
 	static constexpr u32 cMaxDisplayW{ 256 };
 	static constexpr u32 cMaxDisplayH{ 192 };
