+++ src/Systems/CHIP8/Cores/SCHIP_LEGACY.hpp
@@ -15,17 +15,17 @@
 /*==================================================================*/
 
 class SCHIP_LEGACY final : public Chip8_CoreInterface {
-	static constexpr u32 cTotalMemory{ KiB(4) };
+	static constexpr u64 cTotalMemory{ KiB(4) };
 	static constexpr u32 cSafezoneOOB{    32 };
 	static constexpr u32 cGameLoadPos{   512 };
 	static constexpr u32 cStartOffset{   512 };
 	static constexpr f32 cRefreshRate{ 64.0f };
 
-	static constexpr s32 cResSizeMult{     4 };
-	static constexpr s32 cDisplayResW{   128 };
-	static constexpr s32 cDisplayResH{    64 };
-	static constexpr s32 cInstSpeedHi{    45 };
-	static constexpr s32 cInstSpeedLo{    32 };
+	static constexpr s32 cResSizeMult{   4 };
+	static constexpr s32 cDisplayResW{ 128 };
+	static constexpr s32 cDisplayResH{  64 };
+	static constexpr s32 cInstSpeedHi{  45 };
+	static constexpr s32 cInstSpeedLo{  32 };
 
 /*==================================================================*/
